insere Optional nos metodos de AnimalService e ResponseEntity no AnimalController

+++ src/main/java/com/fazenda/services/AnimalService.java
@@ -1,7 +1,7 @@
 package com.fazenda.services;
 
-import java.time.LocalDate;
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -9,42 +9,37 @@ import org.springframework.transaction.annotation.Transactional;
 
 import com.fazenda.dto.AnimalDTO;
 import com.fazenda.dto.AnimalRacaDTO;
-import com.fazenda.dto.AnimalVacinaDTO;
 import com.fazenda.entities.Animal;
-import com.fazenda.entities.AnimalVacina;
-import com.fazenda.entities.AnimalVacinaId;
 import com.fazenda.entities.Categoria;
 import com.fazenda.entities.Raca;
-import com.fazenda.entities.Vacina;
 import com.fazenda.enums.Sexo;
 import com.fazenda.repositories.AnimalRepository;
-import com.fazenda.repositories.AnimalVacinaRepository;
 import com.fazenda.repositories.CategoriaRepository;
 import com.fazenda.repositories.RacaRepository;
-import com.fazenda.repositories.VacinaRepository;
 
 @Service
 public class AnimalService {
 
 	@Autowired
 	private AnimalRepository animalRepository;
-	
+
 	@Autowired
 	RacaRepository racaRepository;
-	
+
 	@Autowired
 	CategoriaRepository categoriaRepository;
 
 	@Transactional(readOnly = true)
 	public List<AnimalDTO> findAll() {
-		List<Animal> animal = animalRepository.findAll();
-		return animal.stream().map(x -> new AnimalDTO(x)).toList();
+		return animalRepository.findAll().stream().map(x -> new AnimalDTO(x)).toList();
+	
 	}
 
 	@Transactional(readOnly = true)
-	public AnimalDTO findById(Long id) {
-		Animal animal = animalRepository.findById(id).get();
-		return new AnimalDTO(animal);
+	public Optional<AnimalDTO> findById(Long id) {
+		Optional<Animal> x = animalRepository.findById(id);
+
+		return x.map(animal -> new AnimalDTO(animal));
 
 	}
 
@@ -56,33 +51,49 @@ public class AnimalService {
 	}
 
 	@Transactional
-	public Animal cadastrarAnimal(Long idRaca, Long idCategoria, Sexo sexo, int meses, double peso) {
-		Animal addAnimal = new Animal();
-		Raca raca = racaRepository.findById(idRaca).get();
-		Categoria categoria = categoriaRepository.findById(idCategoria).get();
-		
-		addAnimal.setCategoria(categoria);
-		addAnimal.setRaca(raca);
-		addAnimal.setSexo(sexo);
-		addAnimal.setPeso(peso);
-		addAnimal.setMeses(meses);
-		
-		return animalRepository.save(addAnimal);
+	public Optional<Animal> cadastrarAnimal(Long idRaca, Long idCategoria, Sexo sexo, int meses, double peso) {
+		Optional<Raca> raca = racaRepository.findById(idRaca);
+		Optional<Categoria> categoria = categoriaRepository.findById(idCategoria);
+
+		if (raca.isPresent() && categoria.isPresent()) {
+			Animal addAnimal = new Animal();
+			addAnimal.setCategoria(categoria.get());
+			addAnimal.setRaca(raca.get());
+			addAnimal.setSexo(sexo);
+			addAnimal.setPeso(peso);
+			addAnimal.setMeses(meses);
+			
+			Animal cadastrado = animalRepository.save(addAnimal);
+			
+			return Optional.of(cadastrado);
+		}
+
+		return Optional.empty();
 	}
 
 	@Transactional
-	public Animal atualizaAnimal(Long id, Long idCategoria, int meses, double peso) {
-		Animal attAnimal = animalRepository.findById(id).get();
-		Categoria categoria = categoriaRepository.findById(idCategoria).get();
-		attAnimal.setMeses(meses);
-		attAnimal.setPeso(peso);
-		attAnimal.setCategoria(categoria);
+	public Optional<Animal> atualizaAnimal(Long id, Long idCategoria, int meses, double peso) {
+		Optional<Animal> attAnimal = animalRepository.findById(id);
+		Optional<Categoria> categoria = categoriaRepository.findById(idCategoria);
 
-		return animalRepository.save(attAnimal);
+		if (attAnimal.isPresent() && categoria.isPresent()) {
+			Animal animal = attAnimal.get();
+			animal.setMeses(meses);
+			animal.setPeso(peso);
+			animal.setCategoria(categoria.get());
+
+			Animal atualizado = animalRepository.save(animal);
+			return Optional.of(atualizado);
+		}
+
+		return Optional.empty();
 	}
 
 	@Transactional
 	public void deleteById(Long id) {
+		if(!animalRepository.existsById(id)) {
+			throw new RuntimeException("Não existe o animal com o id " + id);
+		}
 		animalRepository.deleteById(id);
 	}
 }
